feat: 리뷰 반영 - API 연동 리팩토링
